+++ frontend/src/app/popups/[id]/layout.tsx
@@ -3,5 +3,5 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  return <div className="m-auto w-full p-20 lg:w-1000">{children}</div>;
+  return <div className="m-auto w-full px-20 lg:w-1200">{children}</div>;
 }
